feat(lightbox): support DialogHeading for accessible naming
Split Lightbox into shell + content, provide the ids registry, and derive
its accessible name from a registered DialogHeading.
