+++ packages/lumx-vue/src/components/lightbox/Lightbox.test.tsx
@@ -1,9 +1,12 @@
 import { cleanup, fireEvent, render, screen } from '@testing-library/vue';
 import userEvent from '@testing-library/user-event';
+import { nextTick } from 'vue';
 
 import BaseLightboxTests, { setup } from '@lumx/core/js/components/Lightbox/Tests';
+import { DialogLabelRegistryTests } from '@lumx/core/js/components/Dialog/labelRegistryTests';
 import { CLASSNAME } from '@lumx/core/js/components/Lightbox';
 import { commonTestsSuiteVTL, type SetupRenderOptions } from '@lumx/vue/testing';
+import { DialogHeading } from '@lumx/vue/components/dialog';
 
 import { Lightbox } from '.';
 
@@ -28,6 +31,13 @@ describe('<Lightbox />', () => {
 
     // Core shared tests
     BaseLightboxTests({ render: renderLightbox, screen });
+
+    // Core shared tests: DialogHeading registry wiring.
+    DialogLabelRegistryTests({
+        render: renderLightbox,
+        makeDialogHeading: (name: string) => <DialogHeading>{name}</DialogHeading>,
+        screen,
+    });
 
     const setupLightbox = (props: any = {}, options: SetupRenderOptions<any> = {}) =>
         setup(props, { ...options, render: renderLightbox, screen });
@@ -56,6 +66,26 @@ describe('<Lightbox />', () => {
                 });
                 fireEvent.mouseDown(document.body);
                 expect(emitted('close')).toBeTruthy();
+            });
+        });
+
+        // Registry link/absent cases are covered by the shared DialogLabelRegistryTests above; mechanics
+        // (last-wins, fallback, unmount) by the IdsRegistry tests. Here we only assert Lightbox's own
+        // top-level aria props take precedence over the registry.
+        describe('Label registry', () => {
+            it('should let an explicit aria-label prop override the registry', async () => {
+                render(Lightbox, {
+                    props: {
+                        isOpen: true,
+                        parentElement: document.createElement('div'),
+                        'aria-label': 'Explicit label',
+                    } as any,
+                    slots: { default: () => <DialogHeading>My lightbox</DialogHeading> },
+                });
+                await nextTick();
+                const dialog = screen.getByRole('dialog');
+                expect(dialog).toHaveAttribute('aria-label', 'Explicit label');
+                expect(dialog).not.toHaveAttribute('aria-labelledby');
             });
         });
     });
